feat: Update registration form endpoint to use HTTPS
This commit updates the registration form endpoint to use HTTPS instead of HTTP. This change ensures that the form data is securely transmitted over the network. The previous endpoint "http://localhost:5000/api/competitors" is replaced with "https://jutsa-api.vercel.app/api/competitors/". After submitting the form, a successful registration toast message is displayed to the user.