Add ID field to RegisterForm component
This commit adds a new ID field to the RegisterForm component in order to capture the user's ID number. The ID field is added as an input element with appropriate styling and validation. This change enhances the functionality of the RegisterForm component by including the ID field in the form submission process.